Add task #0 v2

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -8,7 +8,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-	unsigned long int counter;
+	size_t counter;
 
 	while (h != NULL)
 	{
